requirements.txt
Flask
Flask-Cors
langchain-ollama
langchain-core
elevenlabs
GPUtil
psutil
PyPDF2
Pillow
pytesseract